Start by moving test for utils
